+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Angel of Vengeance restores all ability charges over 10 minutes
 - Make creeper explosions not affect tile entities, not destroy dropped items and always drop the blocks broken
 - Save and restore player experience on respawn
+- Core Mod: Allow players to eat food, even when full
 
 ## [0.0.3] - 2020-01-01
 - Initial release of Kittify for Minecraft 1.12.2