+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Make creeper explosions not affect tile entities, not destroy dropped items and always drop the blocks broken
 - Save and restore player experience on respawn
 - Core Mod: Allow players to eat food, even when full
+- Core Mod: Replace regen with own system (1 minute after taking damage, regenerate to full after 5 minutes no matter the health)
 
 ## [0.0.3] - 2020-01-01
 - Initial release of Kittify for Minecraft 1.12.2